Refactor CaffeIO API
This commit refactors the CaffeIO API to align more closely with Caffe.
Now, loading a model is entirely separate from loading the weights, and
making predictions with a pretrained model requires you to load both the
model's .prototxt file and the weights' .caffemodel file.

As of this commit, the CaffeIO workflow is:
1. Create a learner with Net.learner
2. CaffeModel.loadModel(modelFile, net)
3. If you want to load weights: CaffeModel.loadWeights(weightsFile, net)
4. If model is untrained/if you want to fine-tune: learner.train
5. Get predictor from CaffeModel.predictor
6. Predict